move electron update logic into own file, tidy tidy tidy
